+++ api/src/main/java/com/cesi/bhs/api/db/Order.java
@@ -1,0 +1,30 @@
+package com.cesi.bhs.api.db;
+
+import java.sql.Date;
+import java.util.HashMap;
+
+/**
+ *  Interface for Order model, associated with client and products models
+ */
+public interface Order {
+
+  public int getId() ;
+
+  public void setId(int id) ;
+
+  public Date getOrderdate();
+
+  public void setOrderdate(Date orderdate);
+
+  public Date getDeliverydate();
+
+  public void setDeliverydate(Date deliverydate);
+
+  public int getPrice() ;
+
+  public void setPrice(int price);
+
+  public HashMap<String, Integer> getProduct();
+
+  public void setProduct(HashMap<String, Integer> product);
+}
